Solution for the final exercise in the Texture chapter/page
This was a bit of a doozy since I'm not comfortable with pointers in c++. The funny thing is GPT offered a solution that didn't even need to use a pointer by just initializing mixValVal to the value of the original set, which meant the two tracked together! In the end GPT explained pointers to me some (I've asked about them in the past as well) and I now have a better understanding of them. I'm still not super duper comfortable with them, but I'm getting better at understanding what they mean and when to use them.

Code.cpp
- processInput() was changed some to deal with the input and now returns a float based on that input
- Some variables are set before the render loop which are used to update the texture object stuff
- There is an if statement in the redner loop to make sure changes to the mixVal wouldn't go beyond the 0.0-1.0 bounds; if not, the change is made
 - Though during testing I had badly messed up the logic and found out you can safely go outside the 0.0-1.0 bounds
 - Instead of being an issue, it just causes the texture to get "inversely mixed", which looks weird but cool