Go back to the pre-exercise state of the project
The exercise was good to do but it messed with significant pieces of the code that I want to go back to. Luckily I can just go to github and look at what I did for that exercise if I want!

I also forgot to mention in the last commit, fragmentShader.vert had 2 changes
- Add uniform float mixVal; to the shader
- Change the mix equation to use mixVal instead of 0.2f

For this commit
ctrl+z everything I did for exercise 4 until I got to the point before I had started it