Add some mistake handling in doAllTransformations and defined a using for glm2DArray
Rotation requires an input of the rotation angle and rotation axes, meaning the number of the two need to be the same. There is a small check to ensure they are.

The pseudo 2D array I made is now a using statement defined before main; this was to make the code cleaner as to what is being referenced instead of a bunch of weird values.

Code.cpp
- Added an element equality check to doAllTransformations() to ensure that rotations have the same number of elements for each call of the rotate() function
 - If they don't an error message is sent, but the program continues running

- glm2DArray is now the name of the data structure used for the various pieces of the transformation process
 - using glm2DArray = std::vector<glm::vec3>;