First pass at 3D rendering
I added the array for more cubes to be drawn but haven't done that yet as I wanted to get this code saved. Also this doesn't render nicely because I am using elements to draw the cubes rather than vertices, which makes the texture mess up because it is attached to the vertices which were designed for a single face.

Code.cpp
- Added data structues for the cubes for vertices, indices, and model positions
 - These also use their own buffers
- Enabled depth testing
- Clear the buffer bit per frame
- Add some more rotation so I can see each cube face
- Adjusted the draw call for the cube

vertexShader.vert
- Changed GLposition so it also uses the transformation stuff so I can see it rotating